% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{create_conn}
\alias{create_conn}
\title{create a DB Conn}
\usage{
create_conn(param_list, driver = NULL)
}
\arguments{
\item{param_list}{A list of values for creating a pool connection. Would actually prefer passing a config file data directly.}

\item{driver}{A custom driver to connect to database if the need arises}
}
\value{
db connection object
}
\description{
This function just returns a db conn object by passing a config file data to it. You may optionally pass all the parameters as list too.
}
