% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{execute a SQL file}
\usage{
execute_sql_file(sql_conn, sql_file_path, query_params, method = "get")
}
\arguments{
\item{sql_conn}{a connection object be it a pool or a normal connection to the DB}

\item{sql_file_path}{path to an SQL file which has a query}

\item{query_params}{A list of values for interpolation in the SQL file}

\item{method}{only 2 options 'get' or 'post' to either get the data from SQL or just execute a query on the DB server.}
}
\value{
query object
}
\description{
This function runs a .SQL file against a db connection
}
