% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{get_sql_query}
\alias{get_sql_query}
\title{get SQL query object}
\usage{
get_sql_query(sql_conn, sql_file_path, query_params = list())
}
\arguments{
\item{sql_conn}{a connection object be it a pool or a normal connection to the DB}

\item{sql_file_path}{path to an SQL file which has a query}

\item{query_params}{A list of values for interpolation in the SQL file}
}
\value{
query object
}
\description{
This function just returns a db query object from a SQL file path.
}
