% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{read_sql}
\alias{read_sql}
\title{read a SQL file}
\usage{
read_sql(filepath)
}
\arguments{
\item{filepath}{path to an SQL file which has a query}
}
\value{
character string
}
\description{
This function just returns a character string from a .SQL file.
}
