% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{rs_create_pool}
\alias{rs_create_pool}
\title{create a DB pool}
\usage{
rs_create_pool(driver = NULL, param_list)
}
\arguments{
\item{driver}{A custom driver to connect to database if the need arises}

\item{param_list}{A list of values for creating a pool connection. Would actually prefer passing a config file data directly.}

\item{min}{A minimum number of connection for pool object to hold at a point in time}

\item{max}{A maximum number of connection for pool object to hold at a point in time}

\item{idle}{The number of seconds that an idle object will be kept in the pool before it is destroyed (only applies if the number of objects is over the minSize). Use Inf if you want created objects never to be destroyed (there isn't a great reason for this usually).}

\item{...}{Additional arguments to be passed to pool::dbPool function}
}
\value{
pool object
}
\description{
This function just returns a db pool object by passing a config file data to it. You may optionally pass all the parameters as list too.
}
