% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_get_sql_query}
\alias{rs_get_sql_query}
\title{return just SQL text}
\usage{
rs_get_sql_query(sql_query)
}
\arguments{
\item{sql_query}{an Object of type sql_query}
}
\value{
character string
}
\description{
This function just returns the SQL text from a sql_query object
}
