% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_read_query}
\alias{rs_read_query}
\title{read a SQL file}
\usage{
rs_read_query(filepath, sql_query_str = "", method = "dbGetQuery")
}
\arguments{
\item{filepath}{Path to an SQL file containing a query.}

\item{sql_query_str}{A proper SQL statement string. If provided, this is used instead of reading a file.}

\item{method}{The name of the function to execute the query (e.g., "DBI::dbGetQuery", "read_adbc").
Defaults to "dbGetQuery".}
}
\value{
An object of class \code{sql_query} containing the SQL text and execution method.
}
\description{
Creates an object of class \code{sql_query} by reading an SQL file or accepting a raw string.
This object stores the SQL text and the execution method (e.g., "dbGetQuery").
}
