% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{meta_sql_interpolate}
\alias{meta_sql_interpolate}
\title{interpolate meta data in sql query}
\usage{
meta_sql_interpolate(sql_query, meta_query_params)
}
\arguments{
\item{sql_query}{a sql_query object that will be used for sqlinterpolation}

\item{meta_query_params}{A list of values for interpolation in the SQL file
it can also deal with parameter where the we need to build an IN value in sql}
}
\value{
query object
}
\description{
This function just returns a new sql_query object after an interpolation of meta data
}
