% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{print.sql_query}
\alias{print.sql_query}
\title{Prints a sql.query class}
\usage{
\method{print}{sql_query}(x, ...)
}
\arguments{
\item{x}{an Object of type sql_query}
}
\value{
character string
}
\description{
This method just prints a sql_query class nicely
}
