% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_execute}
\alias{rs_execute}
\title{execute a SQL file}
\usage{
rs_execute(sql_query, sql_conn)
}
\arguments{
\item{sql_query}{a sql_query object that will be used for sqlinterpolation}

\item{sql_conn}{a connection object be it a pool or a normal connection to the DB}

\item{query_params}{A list of values for interpolation in the SQL file}
}
\value{
query object
}
\description{
This function runs a .SQL file against a db connection
}
