% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_get_sql_query}
\alias{rs_get_sql_query}
\title{return just SQL text}
\usage{
rs_get_sql_query(sql_query)
}
\arguments{
\item{sql_query}{An object of class \code{sql_query}.}
}
\value{
A character string containing the SQL query.
}
\description{
Extracts and returns the raw SQL text string from a \code{sql_query} object.
}
