% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_interpolate}
\alias{rs_interpolate}
\title{get SQL query object}
\usage{
rs_interpolate(sql_query, sql_conn, query_params = list())
}
\arguments{
\item{sql_query}{a sql_query object that will be used for sqlinterpolation}

\item{sql_conn}{a connection object be it a pool or a normal connection to the DB}

\item{query_params}{A list of values for interpolation in the SQL file}
}
\value{
query object
}
\description{
This function just returns a new sql_query object after interpolation
}
