% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{rs_migrate}
\alias{rs_migrate}
\title{This function runs migration scripts}
\usage{
rs_migrate(sql_conn, up = TRUE)
}
\arguments{
\item{sql_conn}{a sql connection on which all the queries will be executed}

\item{up}{a boolean value to determine if you need to run files from the folder \code{up} or \code{down}
remember these are the only 2 options available.}
}
\value{
a data table with the list of filenames and the rows changed after appending the value
}
\description{
This function runs all the files in the folder sql/migrate/up or sql/migrate/down.
If you want to change the folder set the environment variable \strong{rs_migrate_folder} to a different folder name.
}
