% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{rs_read_query}
\alias{rs_read_query}
\title{read a SQL file}
\usage{
rs_read_query(filepath, method = "get")
}
\arguments{
\item{filepath}{path to an SQL file which has a query}

\item{method}{Only 2 methods are allowed get and post short for getQuery and sendstatement}
}
\value{
character string
}
\description{
This function creates an object of type sql_query by reading the sql file.
}
